requirements.txt
asgiref==3.9.1
Django==5.2.6
django-tailwind==4.4.1
djangorestframework==3.16.1
pillow==11.3.0
pytailwindcss==0.3.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.1
django-crontab
PyJWT